pyproject.toml
[project]
name = "zivv-agent-py"
version = "0.1.0"
description = "Zivv Distributed Agent Worker"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.0.1",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.1",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
package = false
